[FEAT/#18] modal sheet의 view를 상태에 맞게 전환, modal sheet의 크기에 따라 컨텐츠 달라지게 수정
